+++ src/exos/day_01/helperDayOne.ts
