+++ src/exos/day_01/inputs.ts
