+++ src/exos/day_02/helperDayTwo.ts
