+++ src/exos/day_02/inputs.ts
